requirements.txt
pillow
pywin32
mss
screeninfo
pyinstaller